backend/requirements.txt
asgiref==3.11.1
asttokens==3.0.1
decorator==5.2.1
Django==5.1.4
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.2.1
ipython==8.30.0
jedi==0.19.2
matplotlib-inline==0.2.1
parso==0.8.6
pexpect==4.9.0
pillow==12.2.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.20.0
PyJWT==2.12.1
python-decouple==3.8
sqlparse==0.5.5
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.6.0
